requirements.txt
amqp==5.2.0
asgiref==3.8.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
docker==7.0.0
Faker==12.0.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
mixer==7.2.2
packaging==24.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
